packages/parqcast-lambda/pyproject.toml
[project]
name = "parqcast-lambda"
version = "0.1.0"
description = "AWS Lambda handler for parqcast export"
requires-python = "==3.12.3"
license = "Apache-2.0"
dependencies = [
    "parqcast-collectors>=0.1.0",
    "parqcast-transport-s3>=0.1.0",
    "psycopg2-binary>=2.9.11",
]

[tool.uv.sources]
parqcast-transport-s3 = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parqcast"]
